fix:  [Performance] Move search bulk actions logic to a custom hook
